synthrec-core/src/test/java/com/github/steveash/synthrec/canonical/GeneralNormalizerTest.java
/*
 * Copyright (c) 2017, Steve Ash
 *
 * This file is part of Syngen.
 * Syngen is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * Syngen is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with Syngen.  If not, see <http://www.gnu.org/licenses/>.
 */

package com.github.steveash.synthrec.canonical;

import static org.junit.Assert.*;
import static com.google.common.truth.Truth.assertThat;
import static com.google.common.truth.Truth.assertWithMessage;

import org.junit.Test;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

/**
 * @author Steve Ash
 */
public class GeneralNormalizerTest {
    private static final Logger log = LoggerFactory.getLogger(GeneralNormalizerTest.class);

    @Test
    public void shouldNormalizeAddressLine() throws Exception {
        String result = GeneralNormalizer.INSTANCE.normalize("123 West 34th ST N, Dallas, Tx");
        log.info("got: " + result);
        assertThat(result).isEqualTo("123 WEST 34TH ST N, DALLAS, TX");

    }
}